refactor(console): apply no-unused-modules
