+++ packages/console/package.json
@@ -142,8 +142,22 @@
           "namedComponents": "function-declaration",
           "unnamedComponents": "arrow-function"
         }
+      ],
+      "import/no-unused-modules": [
+        "error",
+        { "unusedExports": true }
       ]
-    }
+    },
+    "overrides": [
+      {
+        "files": [
+          "*.d.ts", "**/mdx-components/*/index.tsx"
+        ],
+        "rules": {
+          "import/no-unused-modules": "off"
+        }
+      }
+    ]
   },
   "stylelint": {
     "extends": "@silverhand/eslint-config-react/.stylelintrc"
