--- packages/console/src/components/Uploader/index.tsx
@@ -1,3 +1,0 @@
-export { default as FileUploader } from './FileUploader';
-export { default as ImageUploader } from './ImageUploader';
-export { default as ImageUploaderField } from './ImageUploaderField';
