+++ packages/console/src/hooks/use-confirm-modal.ts
@@ -2,6 +2,4 @@ import { useContext } from 'react';
 
 import { AppConfirmModalContext } from '@/contexts/AppConfirmModalProvider';
 
-export type { ModalContentRenderProps } from '@/contexts/AppConfirmModalProvider';
-
 export const useConfirmModal = () => useContext(AppConfirmModalContext);
